refactor: Move MemeRenderer and MemeTextPainter to editor package
- Moved `MemeRenderer` and `MemeTextPainter` from the `domain` package to the `editor` package within the presentation layer.
- This change better reflects the responsibility of these classes, which are primarily used for rendering and drawing text on memes within the editor.
- Updated imports and code references to reflect the new package location.
